backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
duckduckgo-search==4.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
authlib==1.3.0 